+++ Tech-Challenge/src/main/java/com/fiap/techChallenge/TechChallenge/domain/Eletrodomestico.java
@@ -2,7 +2,6 @@ package com.fiap.techChallenge.TechChallenge.domain;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.stereotype.Component;
 
 import java.util.Random;
 @Getter@Setter
@@ -11,7 +10,7 @@ public class Eletrodomestico {
     private Integer id ;
     private String nome;
     private String modelo;
-    private Integer potencia;
+    private Double potencia;
 
     public void setId(){
         Random random = new Random();
